third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-playback-setmediakeys-before-src.html

<!DOCTYPE html>
<html>
    <head>
        <title>Clear Key Playback</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="testVideo"></video>
        <div id="log"></div>
        <p>Test playback of encrypted media using Clear Key key system.</p>
        <script>
            async_test(function(test)
            {
                var video = document.getElementById('testVideo');
                var isSessionCreated = false;
                var isUpdatePromiseResolved = false;

                var rawKey = new Uint8Array([0xeb, 0xdd, 0x62, 0xf1, 0x68, 0x14, 0xd2, 0x7b,
                                             0x68, 0xef, 0x12, 0x2a, 0xfc, 0xe4, 0xae, 0x3c]);

                function onNeedKey(event)
                {
                    assert_equals(event.target, video);
                    assert_true(event instanceof window.MediaKeyNeededEvent);
                    assert_equals(event.type, 'needkey');

                    // The same decryption key is shared by all streams so only create a shared session once.
                    if (isSessionCreated)
                        return;
                    isSessionCreated = true;

                    var mediaKeySession = video.mediaKeys.createSession();
                    waitForEventAndRunStep('message', mediaKeySession, onMessage, test);
                    mediaKeySession.generateRequest(event.contentType, event.initData).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                }

                function onMessage(event)
                {
                    assert_true(event instanceof window.MediaKeyMessageEvent);
                    assert_equals(event.type, 'message');

                    var keyId = extractSingleKeyIdFromMessage(event.message);
                    var jwkSet = stringToUint8Array(createJWKSet(createJWK(keyId, rawKey)));
                    event.target.update(jwkSet).then(function(result) {
                        isUpdatePromiseResolved = true;
                    }).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                }

                function onPlaying(event)
                {
                    // Not using waitForEventAndRunStep() to avoid too many
                    // EVENT(onTimeUpdate) logs.
                    video.addEventListener('timeupdate', onTimeUpdate, true);
                }

                function onTimeUpdate(event)
                {
                    if (event.target.currentTime < 0.2 || !isUpdatePromiseResolved)
                        return;

                    test.done();
                }

                MediaKeys.create('org.w3.clearkey').then(function(mediaKeys) {

                    waitForEventAndRunStep('needkey', video, onNeedKey, test);
                    waitForEventAndRunStep('playing', video, onPlaying, test);

                    video.setMediaKeys(mediaKeys);
                    video.src = '../content/test-encrypted.webm';
                    video.play();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'Playback using Clear Key key system, calling setMediaKeys() before setting src attribute.');
        </script>
    </body>
</html>
