third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-playback-setmediakeys-before-src-expected.txt

Test playback of encrypted media using Clear Key key system.

EVENT(encrypted)
EVENT(encrypted)
EVENT(message)
EVENT(playing)
This is a testharness.js-based test.
PASS Playback using Clear Key key system, calling setMediaKeys() before setting src attribute. 
Harness: the test ran to completion.
